Week_2/Day5/project/requirements.txt
Flask
Flask-cors
translate
gTTS 
python-dotenv
psycopg
psycopg2-binary
